requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
api==0.0.7
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
distro==1.9.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==65.0.2
django-cors-headers==4.5.0
django-etc==1.4.0
django-hitcount==1.3.5
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
Faker==30.0.0
frozenlist==1.4.1
gevent==24.10.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.2
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.131
msgpack==1.1.0
multidict==6.1.0
nose==1.3.7
numpy==1.26.4
openai==1.51.0
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
tenacity==8.5.0
tiktoken==0.8.0
tomli==2.0.2
toposort==1.10
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
websocket==0.2.1
yarl==1.13.1
zope.event==5.0
zope.interface==7.1.0